dbcleaner: Create utility with first command
Create the first utility (dbcleaner) with one command but allowing to be
extended with more if needed.
Also create some common internal packages.

This commit brings the initial go.mod file which defines the repository,
the Go tooling dependencies and also the dependencies required by the
implementation of this first utility.